Separate durable agent examples by mode
