+++ examples/openai-sdk-agent/agent_recovery_polling/__init__.py
@@ -1,0 +1,1 @@
+"""Agent-session-managed recovery without SSE replay."""
