+++ examples/openai-sdk-agent/agent_recovery_sse/__init__.py
@@ -1,0 +1,1 @@
+"""Agent-session-managed recovery with SSE replay."""
