+++ examples/openai-sdk-agent/agent_recovery_sse/app.yaml
@@ -1,4 +1,4 @@
-command: ["python", "app_agent_recovery_sse.py"]
+command: ["python", "-m", "agent_recovery_sse.app"]
 env:
   - name: OPENAI_API_KEY
     valueFrom: openai-api-key
