+++ examples/openai-sdk-agent/framework_recovery_sse/__init__.py
@@ -1,0 +1,1 @@
+"""Framework-managed recovery with SSE replay."""
