+++ examples/openai-sdk-agent/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared OpenAI Agents SDK implementation for all recovery modes."""
